refactor: move order amount helper to utils
